# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RiskIntel360 Platform API
  description: |
    RiskIntel360 is an autonomous multi-agent business intelligence platform that leverages Amazon Bedrock Nova 
    and AWS AI services to revolutionize enterprise market validation. The platform transforms traditional 
    manual market research (taking weeks) into intelligent, automated workflows that deliver comprehensive 
    business validation insights in under 2 hours.
    
    ## Key Features
    - **Multi-Agent Architecture**: Six specialized AI agents working in coordination
    - **Real-time Processing**: Complete validation workflows in under 2 hours
    - **Enterprise Security**: OAuth 2.0 authentication with role-based access control
    - **Scalable Infrastructure**: Auto-scaling from 3-50 instances based on demand
    
    ## Agent Types
    - **Market Research Agent**: Market data aggregation and trend analysis
    - **Competitive Intelligence Agent**: Competitor analysis and positioning assessment
    - **Financial Validation Agent**: Revenue projections and investment analysis
    - **Risk Assessment Agent**: Business risk evaluation and compliance analysis
    - **Customer Intelligence Agent**: Customer behavior analysis and market demand validation
    - **Synthesis & Recommendation Agent**: Cross-agent insight integration and strategic recommendations
  version: 1.0.0
  contact:
    name: RiskIntel360 Platform Team
    email: support@RiskIntel360.com
  license:
    name: MIT
...
servers:
...
security:
  - BearerAuth: []
  - OAuth2: [read, write]
paths:
...
  /auth/login:
    post:
      summary: User authentication
      description: Authenticate user with OAuth 2.0 credentials
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: Refresh authentication token
      description: Refresh an expired access token using a refresh token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.RiskIntel360.com/oauth/authorize
          tokenUrl: https://auth.RiskIntel360.com/oauth/token
          scopes:
            read: Read access to validation data
            write: Create and modify validation requests
            admin: Administrative access
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          description: Unique request identifier for tracking

    # Additional schemas for completeness
...